styles(Nav): Move nav styles into its folder
Breaks the styles off from the main stylesheet and into the `Navigation`
component folder; as a consequence, the `postcss.config.js` file must be
moved into the root directory so that it is detected from the context of
component folders. I may use the root configuration exclusively in the
future.